backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.20.2
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter-client==8.0.2
jupyter-core==5.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.1
openpyxl==3.1.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
PySocks==1.7.1
pysondb-v2==2.1.0
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
requests==2.28.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.2
zipp==3.12.0
zope.event==4.6
zope.interface==5.5.2
